save chromosome in a new format
